requirements.txt
django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
reportlab==4.4.5
networkx==3.5
bibx==0.9.0